backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
clerk-backend-sdk==1.0.0
recipe-scrapers==14.54.0
Pillow==10.1.0
google-cloud-vision==3.4.5
openai==1.3.0
pytest==7.4.3
pytest-asyncio==0.21.1